Pass integers to np.linspace/np.logspace's count.
This fixes "DeprecationWarning: object of type <class 'float'> cannot
be safely interpreted as an integer." raised by latest NumPy.

+++ lib/matplotlib/projections/geo.py
@@ -190,7 +190,7 @@ class GeoAxes(Axes):
         """
         Set the number of degrees between each longitude grid.
         """
-        number = (360.0 / degrees) + 1
+        number = int(360 / degrees) + 1
         self.xaxis.set_major_locator(
             FixedLocator(
                 np.linspace(-np.pi, np.pi, number, True)[1:-1]))
@@ -200,7 +200,7 @@ class GeoAxes(Axes):
         """
         Set the number of degrees between each longitude grid.
         """
-        number = (180.0 / degrees) + 1
+        number = int(180 / degrees) + 1
         self.yaxis.set_major_locator(
             FixedLocator(
                 np.linspace(-np.pi / 2.0, np.pi / 2.0, number, True)[1:-1]))
